+++ feature/metis/manage-conversations/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/manageconversations/ui/conversation/overview/ConversationOverviewViewModel.kt
@@ -432,16 +432,4 @@ class ConversationOverviewViewModel(
         isExpanded: Boolean,
         showPrefix: Boolean = true
     ) = ConversationCollection(this, isExpanded, showPrefix)
-
-    fun removeSectionPrefix(value: String): String {
-        val prefixes = listOf("exercise-", "lecture-", "exam-")
-        var result = value
-        for (prefix in prefixes) {
-            if (result.startsWith(prefix, ignoreCase = true)) {
-                result = result.removePrefix(prefix)
-                break
-            }
-        }
-        return result.trim()
-    }
 }
